+++ Simulator/Edges/Neuro/AllSpikingSynapses_d.cpp
@@ -405,7 +405,7 @@ void AllSpikingSynapses::printGPUEdgesProps(void *allEdgesDeviceProps) const
             cout << "GPU W[" << i << "] = " << WPrint[i];
             cout << " GPU sourNeuron: " << sourceNeuronIndexPrint[i];
             cout << " GPU desNeuron: " << destNeuronIndexPrint[i];
-            cout << " GPU type: " << static_cast<int>(typePrint[i]);
+            cout << " GPU type: " << typePrint[i];
             cout << " GPU psr: " << psrPrint[i];
             cout << " GPU in_use:" << (inUsePrint[i] == 1 ? "true" : "false");
             cout << " GPU decay: " << decayPrint[i];
